requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlmodel==0.0.16
httpx==0.27.0
joblib==1.4.0
scikit-learn==1.5.0
python-dotenv==1.0.0